Replace Taskable to TaskLike
