--- shared/src/main/scala/io/casperlabs/catscontrib/TaskableOps.scala
@@ -1,12 +1,0 @@
-package io.casperlabs.catscontrib
-
-import monix.eval.Task
-
-final class TaskableOps[F[_]: Taskable, A](fa: F[A]) {
-  def toTask: Task[A] = Taskable[F].toTask(fa)
-}
-
-trait ToTaskableOps {
-  implicit def ToTaskableOps[F[_]: Taskable, A](fa: F[A]): TaskableOps[F, A] =
-    new TaskableOps[F, A](fa)
-}
